+++ AllaganLib.GameSheets/LuminaSheets/QuestDialogue.cs
@@ -1,0 +1,22 @@
+using Lumina.Excel;
+using Lumina.Text.ReadOnly;
+
+namespace AllaganLib.GameSheets.LuminaSheets;
+
+[Sheet("QuestDialogue")]
+public readonly struct QuestDialogue(RawRow row) : IExcelRow<QuestDialogue>
+{
+    public uint RowId => row.RowId;
+
+    public ReadOnlySeString Key => row.ReadStringColumn(0);
+
+    public ReadOnlySeString Value => row.ReadStringColumn(1);
+
+    public ExcelPage ExcelPage => row.ExcelPage;
+    public uint RowOffset => row.RowOffset;
+
+    static QuestDialogue IExcelRow<QuestDialogue>.Create(ExcelPage page, uint offset, uint row)
+    {
+        return new QuestDialogue(new RawRow(page, offset, row));
+    }
+}