+++ AllaganLib.GameSheets/Sheets/ClassJobSheet.cs
@@ -9,9 +9,16 @@ namespace AllaganLib.GameSheets.Sheets;
 
 public class ClassJobSheet : ExtendedSheet<ClassJob, ClassJobRow, ClassJobSheet>, IExtendedSheet
 {
+    private ClassJobCategorySheet? classJobCategorySheet;
+
     public ClassJobSheet(GameData gameData, SheetManager sheetManager, SheetIndexer sheetIndexer, ItemInfoCache itemInfoCache)
         : base(gameData, sheetManager, sheetIndexer, itemInfoCache)
     {
+    }
+
+    public ClassJobCategorySheet GetClassCategorySheet()
+    {
+        return classJobCategorySheet ??= this.SheetManager.GetSheet<ClassJobCategorySheet>();
     }
 
     public override void CalculateLookups()