+++ AllaganLib.GameSheets/Sheets/Rows/MapRow.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using AllaganLib.GameSheets.Model;
 using Lumina.Excel.Sheets;
 
@@ -13,15 +14,32 @@ public class MapRow : ExtendedRow<Map, MapRow, MapSheet>
         {
             if (this.formattedName == null)
             {
-                var map = this.Base.PlaceName.ValueNullable?.Name.ExtractText() ?? "Unknown Map";
-                var region = this.Base.PlaceNameRegion.ValueNullable?.Name.ToString() ?? "Unknown Territory";
+                var map = this.Base.PlaceName.ValueNullable?.Name.ExtractText() ?? null;
+                var region = this.Base.PlaceNameRegion.ValueNullable?.Name.ToString() ?? null;
                 var subArea = this.Base.PlaceNameSub.ValueNullable?.Name.ToString() ?? null;
-                if (!string.IsNullOrEmpty(subArea))
+
+                if (region == "???")
                 {
-                    subArea = " - " + subArea;
+                    region = null;
                 }
 
-                this.formattedName = region + " - " + map + (subArea ?? string.Empty);
+                var parts = new List<string>();
+                if (!string.IsNullOrEmpty(region))
+                {
+                    parts.Add(region);
+                }
+
+                if (!string.IsNullOrEmpty(map))
+                {
+                    parts.Add(map);
+                }
+
+                if (!string.IsNullOrEmpty(subArea))
+                {
+                    parts.Add(subArea);
+                }
+
+                this.formattedName = string.Join(" - ", parts);
             }
 
             return this.formattedName;