+++ CHANGELOG.md
@@ -13,6 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [2.1.0] - 2026-03-09
+### Added
+- AllaganLib.GameSheets
+  - Added SharedModelCache
+  - Added RoleType to ClassJobRow
+  - Added QuestDialogue sheet
+
+- AllaganLib.Monitors
+  - ShopMonitorService now provides the list of shops, subshops and active shop when visiting a vendor with a nested menu(topic select)
+
 ## [2.0.6] - 2026-02-07
 ### Added
 - AllaganLib.GameSheets